fix:#29 key 변경 + rememberLauncher 메소드 사용
