+++ app/src/main/java/com/dev/briefing/presentation/login/SignInScreen.kt
@@ -1,4 +1,10 @@
 import android.content.Intent
+import android.util.Log
+import android.widget.Toast
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -23,12 +29,15 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat.startActivity
+import com.dev.briefing.BuildConfig
 import com.dev.briefing.BuildConfig.GOOGLE_API_KEY
 import com.dev.briefing.R
 import com.dev.briefing.presentation.home.HomeActivity
+import com.dev.briefing.presentation.login.mGoogleSignInClient
 import com.dev.briefing.presentation.theme.MainPrimary
 import com.dev.briefing.presentation.theme.MainPrimary2
 import com.dev.briefing.presentation.theme.MainPrimary3
@@ -36,12 +45,16 @@ import com.dev.briefing.presentation.theme.MainPrimary5
 import com.dev.briefing.presentation.theme.Typography
 import com.dev.briefing.presentation.theme.White
 import com.dev.briefing.presentation.theme.utils.CommonDialog
+import com.google.android.gms.auth.api.Auth.GoogleSignInApi
 import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.android.gms.common.api.ApiException
+import com.google.android.gms.tasks.Task
 
 @Composable
 fun SignInScreen(
-    googelSignIn:() -> Unit = {}
+    googelSignIn: () -> Unit = {}
 ) {
     val context = LocalContext.current
     val openAlertDialog = remember { mutableStateOf(false) }
@@ -84,7 +97,7 @@ fun SignInScreen(
 
             Text(
                 text = stringResource(R.string.login_app_description),
-                style = Typography.titleMedium.copy(color = White)
+                style = Typography.titleMedium.copy(color = White, fontWeight = FontWeight.Normal)
             )
             Spacer(modifier = Modifier.height(19.dp))
             Image(
@@ -115,12 +128,32 @@ fun SignInScreen(
 fun GoogleLoginButton(
     onClick: () -> Unit = {}
 ) {
+    //4. startActivityForResult를 통해 구글 로그인 창을 띄움
+    val googleCallBack =
+        rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
+            Log.d("Google", result.resultCode.toString())
+            // Intent 타입으로 받은 data를 GoogleSignInResult로 바꿔준다.
+            var rs = result.data?.let { GoogleSignInApi.getSignInResultFromIntent(it) }
+            Log.e("RESULT", rs?.status.toString())
+            Log.e("RESULT", result.data.toString())
+            if (result.resultCode == ComponentActivity.RESULT_OK) {
+                val intent = result.data
+                if (result.data != null) {
+                    val task: Task<GoogleSignInAccount> =
+                        GoogleSignIn.getSignedInAccountFromIntent(intent)
+                    handleSignInResult(task)
+                }
+            }
+        }
     Row(
         modifier = Modifier
             .fillMaxWidth()
             .background(color = White, shape = RoundedCornerShape(25.dp))
             .padding(vertical = 13.dp)
-            .clickable(onClick = onClick),
+            .clickable(onClick = {
+                onClick()
+                googleCallBack.launch(mGoogleSignInClient?.signInIntent)
+            }),
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.Center
     ) {
@@ -133,5 +166,29 @@ fun GoogleLoginButton(
             text = "Log in with Google",
             style = Typography.headlineLarge.copy(color = Color(0xFF7C7C7C))
         )
+    }
+}
+
+//5. 사용자 정보 가져오기
+fun handleSignInResult(completedTask: Task<GoogleSignInAccount>) {
+    try {
+        Log.d("Google", completedTask.isSuccessful.toString())
+        val account = completedTask.getResult(ApiException::class.java)
+
+        Log.d("Google", account.account.toString())
+        Log.d("Google", account.displayName.toString())
+        Log.d("Google", account.idToken.toString())
+        Log.d("Google", account.id.toString())
+        account.idToken?.let {
+            Log.d("Google", it)
+        } ?: Log.d("Google", "구글 로그인 에러")
+
+        // Signed in successfully, show authenticated UI.
+//            updateUI(account)
+    } catch (e: ApiException) {
+        // The ApiException status code indicates the detailed failure reason.
+        // Please refer to the GoogleSignInStatusCodes class reference for more information.
+        Log.d("Google", "google 로그인에 실패하였습니다. 다시시도해주세요${e}")
+
     }
 }
